+++ transpiler/__tests__/end-to-end/symbol-table.steps.ts
@@ -22,11 +22,6 @@ import { BitloopsParser } from '../../src/parser/index.js';
 import { IntermediateASTParser } from '../../src/ast/core/index.js';
 import { TargetGenerator } from '../../src/target/index.js';
 import { SYMBOL_TABLE_TEST_CASES } from './mocks/symbol-table/symbol-table.js';
-import { SYMBOL_TABLE_ALREADY_DECLARED_TEST_CASES } from './mocks/symbol-table/already-declared/index.js';
-import { SYMBOL_TABLE_CONSTANT_REASSIGNMENT_TEST_CASES } from './mocks/symbol-table/constant-reassignment/index.js';
-import { SYMBOL_TABLE_MISSING_IDENTIFIERS_TEST_CASES } from './mocks/symbol-table/missing-identifiers/index.js';
-import { SYMBOL_TABLE_MEMBER_NOT_DEFINED_TEST_CASES } from './mocks/symbol-table/not-defined-member/index.js';
-import { SYMBOL_TABLE_FIND_TYPE_OF_KEYWORD_TEST_CASES } from './mocks/symbol-table/type-of-keyword/index.js';
 
 describe('Symbol table cases', () => {
   const boundedContext = 'Hello world';
@@ -74,234 +69,6 @@ describe('Symbol table cases', () => {
       expect(result.symbolTables[boundedContext].getJsonValue()).toEqual(
         testCase.expectedSymbolTable,
       );
-    });
-  });
-});
-
-describe.skip('Validation checks table cases', () => {
-  const boundedContext = 'Hello world';
-  const module = 'demo';
-
-  describe('Missing identifiers test cases', () => {
-    SYMBOL_TABLE_MISSING_IDENTIFIERS_TEST_CASES.forEach((testCase) => {
-      const parser = new BitloopsParser();
-      const originalLanguageASTToIntermediateModelTransformer = new IntermediateASTParser();
-      const intermediateASTModelToTargetLanguageGenerator = new TargetGenerator();
-
-      const transpiler = new Transpiler(
-        parser,
-        originalLanguageASTToIntermediateModelTransformer,
-        intermediateASTModelToTargetLanguageGenerator,
-      );
-
-      it(`${testCase.description}`, async () => {
-        // given
-        const input = {
-          core: [
-            {
-              boundedContext,
-              module,
-              fileId: 'fileId',
-              fileContents: testCase.inputCore,
-            },
-          ],
-          setup: [
-            {
-              boundedContext,
-              module,
-              fileId: 'fileId',
-              fileContents: testCase.inputSetup,
-            },
-          ],
-        };
-
-        // when
-        const result = transpiler.getSymbolTable(input);
-        if (Transpiler.isTranspilerError(result)) {
-          throw new Error('Transpiler should NOT return error');
-        }
-        expect(result.semanticErrors.map((x) => x.message)).toEqual(testCase.errorMessages);
-      });
-    });
-  });
-
-  describe('constant reassingment identifiers test cases', () => {
-    SYMBOL_TABLE_CONSTANT_REASSIGNMENT_TEST_CASES.forEach((testCase) => {
-      const parser = new BitloopsParser();
-      const originalLanguageASTToIntermediateModelTransformer = new IntermediateASTParser();
-      const intermediateASTModelToTargetLanguageGenerator = new TargetGenerator();
-
-      const transpiler = new Transpiler(
-        parser,
-        originalLanguageASTToIntermediateModelTransformer,
-        intermediateASTModelToTargetLanguageGenerator,
-      );
-
-      it(`${testCase.description}`, async () => {
-        // given
-        const input = {
-          core: [
-            {
-              boundedContext,
-              module,
-              fileId: 'fileId',
-              fileContents: testCase.inputCore,
-            },
-          ],
-          setup: [
-            {
-              boundedContext,
-              module,
-              fileId: 'fileId',
-              fileContents: testCase.inputSetup,
-            },
-          ],
-        };
-
-        // when
-        const result = transpiler.getSymbolTable(input);
-        if (Transpiler.isTranspilerError(result)) {
-          throw new Error('Transpiler should NOT return error');
-        }
-        expect(result.semanticErrors.map((x) => x.message)).toEqual(testCase.errorMessages);
-      });
-    });
-  });
-
-  describe('Already declared variables test cases', () => {
-    SYMBOL_TABLE_ALREADY_DECLARED_TEST_CASES.forEach((testCase) => {
-      const parser = new BitloopsParser();
-      const originalLanguageASTToIntermediateModelTransformer = new IntermediateASTParser();
-      const intermediateASTModelToTargetLanguageGenerator = new TargetGenerator();
-
-      const transpiler = new Transpiler(
-        parser,
-        originalLanguageASTToIntermediateModelTransformer,
-        intermediateASTModelToTargetLanguageGenerator,
-      );
-
-      it(`${testCase.description}`, async () => {
-        // given
-        const input = {
-          core: [
-            {
-              boundedContext,
-              module,
-              fileId: 'fileId',
-              fileContents: testCase.inputCore,
-            },
-          ],
-          setup: [
-            {
-              boundedContext,
-              module,
-              fileId: 'fileId',
-              fileContents: testCase.inputSetup,
-            },
-          ],
-        };
-
-        // when
-        const result = transpiler.getSymbolTable(input);
-        if (Transpiler.isTranspilerError(result)) {
-          throw new Error('Transpiler should NOT return error');
-        }
-        expect(result.semanticErrors.map((x) => x.message)).toEqual(testCase.errorMessages);
-      });
-    });
-  });
-
-  describe('Not defined members test cases', () => {
-    SYMBOL_TABLE_MEMBER_NOT_DEFINED_TEST_CASES.forEach((testCase) => {
-      const parser = new BitloopsParser();
-      const originalLanguageASTToIntermediateModelTransformer = new IntermediateASTParser();
-      const intermediateASTModelToTargetLanguageGenerator = new TargetGenerator();
-
-      const transpiler = new Transpiler(
-        parser,
-        originalLanguageASTToIntermediateModelTransformer,
-        intermediateASTModelToTargetLanguageGenerator,
-      );
-
-      it(`${testCase.description}`, async () => {
-        // given
-        const input = {
-          core: [
-            {
-              boundedContext,
-              module,
-              fileId: 'fileId',
-              fileContents: testCase.inputCore,
-            },
-          ],
-          setup: [
-            {
-              boundedContext,
-              module,
-              fileId: 'fileId',
-              fileContents: testCase.inputSetup,
-            },
-          ],
-        };
-
-        // when
-        const result = transpiler.getSymbolTable(input);
-        if (Transpiler.isTranspilerError(result)) {
-          throw new Error('Transpiler should NOT return error');
-        }
-        expect(result.semanticErrors.map((x) => x.message)).toEqual(testCase.errorMessages);
-      });
-    });
-  });
-});
-
-describe.skip('Find type of keyword test cases', () => {
-  const boundedContext = 'Hello world';
-  const module = 'demo';
-  SYMBOL_TABLE_FIND_TYPE_OF_KEYWORD_TEST_CASES.forEach((testCase) => {
-    const parser = new BitloopsParser();
-    const originalLanguageASTToIntermediateModelTransformer = new IntermediateASTParser();
-    const intermediateASTModelToTargetLanguageGenerator = new TargetGenerator();
-
-    const transpiler = new Transpiler(
-      parser,
-      originalLanguageASTToIntermediateModelTransformer,
-      intermediateASTModelToTargetLanguageGenerator,
-    );
-
-    it(`${testCase.description}`, async () => {
-      // given
-      const input = {
-        core: [
-          {
-            boundedContext,
-            module,
-            fileId: 'fileId',
-            fileContents: testCase.inputCore,
-          },
-        ],
-        setup: [
-          {
-            boundedContext,
-            module,
-            fileId: 'fileId',
-            fileContents: testCase.inputSetup,
-          },
-        ],
-      };
-
-      // when
-      const result = transpiler.getSymbolTable(input);
-      if (Transpiler.isTranspilerError(result)) {
-        throw new Error('Transpiler should NOT return error');
-      }
-      const symbolTable = result.symbolTables[boundedContext];
-      const output = symbolTable.findTypeOfKeyword(
-        testCase.keywordInfo.name,
-        testCase.keywordInfo.position,
-      );
-
-      expect(output).toEqual(testCase.expectedOutput);
     });
   });
 });
